Add new algorithms and update gitignore
This commit introduces four new algorithm solutions (MountainArray, ThirdMax, DuplicateVal, HeightChecker), and an incomplete solution for DisappearedNums. It also includes updates to .gitignore to ignore files and directories that are not necessary to track. Lastly, a README file was added to ConsoleApp1 for documenting the purpose of these algorithm files.