Refactor graph traversal and add Leetcode863 class
The changes simplify the conditional logic in the existing graph traversal method by utilizing a continue statement. Moreover, a new 'Leetcode863' class has been added, which implements a TreeNode structure and allows for processing of tree-based graph structures via BFS and DFS methods. The methods allow to transform a tree into a graph and to calculate distances to a given target node.