Add minimum mutation function in Leetcode433 class
The Leetcode433 class is introduced to find the minimum number of mutations needed to change a start gene to an end gene from a given bank of genes. Using a breadth-first search, the implemented MinMutation method calculates steps for the transformation. It also includes a Neighbor helper method to find all possible mutations for a given gene.